+++ src/onegov.core/onegov/core/tests/test_converters.py
@@ -1,5 +1,7 @@
-from datetime import date
-from onegov.core.converters import extended_date_converter, uuid_converter
+from datetime import date, datetime
+from onegov.core.converters import extended_date_converter
+from onegov.core.converters import datetime_converter
+from onegov.core.converters import uuid_converter
 from uuid import UUID
 
 
@@ -13,6 +15,19 @@ def test_date_converter():
     assert converter.decode(['']) is None
     assert converter.decode([None]) is None
     assert converter.decode(['2008-12-30']) == date(2008, 12, 30)
+
+
+def test_datetime_converter():
+    converter = datetime_converter
+
+    assert converter.encode(None) == ['']
+    assert converter.encode('') == ['']
+    assert converter.encode(datetime(2008, 12, 30)) == ['2008-12-30T00:00:00']
+
+    assert converter.decode(['']) is None
+    assert converter.decode([None]) is None
+    assert converter.decode(['2008-12-30T12:34:56']) == datetime(
+        2008, 12, 30, 12, 34, 56)
 
 
 def test_uuid_converter():